# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KESHAV",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/risk/current/{location_id}": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Get Current Risk",
        "description": "Get current health risk for a location.\n\nUses most recent thermal observation and vulnerability data.",
        "operationId": "get_current_risk_api_v1_risk_current__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskPredictionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/forecast/{location_id}": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Get Forecast Risk",
        "description": "Get health risk forecast for a location.\n\nUses forecast thermal data for the specified horizon.",
        "operationId": "get_forecast_risk_api_v1_risk_forecast__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "horizon_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "default": 3,
              "title": "Horizon Days"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskForecastOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/history/{location_id}": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Get Risk History",
        "description": "Get historical risk predictions for a location.",
        "operationId": "get_risk_history_api_v1_risk_history__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskPredictionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/calculate": {
      "post": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Calculate Risk",
        "description": "Calculate health risk from custom meteorological inputs.\n\nAllows calculation without requiring existing weather data in the database.",
        "operationId": "calculate_risk_api_v1_risk_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskCalculateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskPredictionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/train": {
      "post": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Train Model",
        "description": "Train a new health risk model.\n\nRequires ADMIN or ML_ENGINEER role.",
        "operationId": "train_model_api_v1_risk_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskTrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskTrainingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/training/{run_id}": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Get Training Run",
        "description": "Get training run status and results.",
        "operationId": "get_training_run_api_v1_risk_training__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskTrainingRunOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/models": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "List Models",
        "description": "List models in registry.",
        "operationId": "list_models_api_v1_risk_models_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "algorithm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Algorithm"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Models Api V1 Risk Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/models/{model_id}": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Get Model",
        "description": "Get model details from registry.",
        "operationId": "get_model_api_v1_risk_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Model Api V1 Risk Models  Model Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/models/{model_id}/promote": {
      "post": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Promote Model Endpoint",
        "description": "Promote a model to production.",
        "operationId": "promote_model_endpoint_api_v1_risk_models__model_id__promote_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Promote Model Endpoint Api V1 Risk Models  Model Id  Promote Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/models/{model_id}/retire": {
      "post": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Retire Model Endpoint",
        "description": "Retire a model.",
        "operationId": "retire_model_endpoint_api_v1_risk_models__model_id__retire_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Retire Model Endpoint Api V1 Risk Models  Model Id  Retire Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/evaluation/{model_id}": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Get Model Evaluation",
        "description": "Get model evaluation metrics.",
        "operationId": "get_model_evaluation_api_v1_risk_evaluation__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "split",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Split"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskEvaluationOutput"
                  },
                  "title": "Response Get Model Evaluation Api V1 Risk Evaluation  Model Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/features": {
      "get": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Get Feature Manifest",
        "description": "Get feature manifest for a version.",
        "operationId": "get_feature_manifest_api_v1_risk_features_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1.0",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Feature Manifest Api V1 Risk Features Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/risk/artifacts/cleanup": {
      "post": {
        "tags": [
          "health-risk",
          "health-risk"
        ],
        "summary": "Cleanup Artifacts",
        "description": "Clean up old model artifacts.",
        "operationId": "cleanup_artifacts_api_v1_risk_artifacts_cleanup_post",
        "parameters": [
          {
            "name": "keep_latest",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10,
              "title": "Keep Latest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cleanup Artifacts Api V1 Risk Artifacts Cleanup Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskForecastOutput": {
        "properties": {
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "prediction_time": {
            "type": "string",
            "format": "date-time",
            "title": "Prediction Time"
          },
          "forecasts": {
            "items": {},
            "type": "array",
            "title": "Forecasts"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "feature_version": {
            "type": "string",
            "title": "Feature Version"
          }
        },
        "type": "object",
        "required": [
          "location_id",
          "prediction_time",
          "forecasts",
          "model_id",
          "model_version",
          "feature_version"
        ],
        "title": "RiskForecastOutput",
        "description": "Forecast risk output for multiple days."
      },
...
      "RiskPredictionList": {
        "properties": {
          "predictions": {
            "items": {},
            "type": "array",
            "title": "Predictions"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "predictions"
        ],
        "title": "RiskPredictionList"
      },
      "RiskPredictionOutput": {
        "properties": {
          "prediction_id": {
            "type": "string",
            "title": "Prediction Id"
          },
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "prediction_time": {
            "type": "string",
            "format": "date-time",
            "title": "Prediction Time"
          },
          "target_date": {
            "type": "string",
            "format": "date-time",
            "title": "Target Date"
          },
          "horizon_days": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Horizon Days",
            "default": 0
          },
          "risk_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Probability"
          },
          "risk_category": {
            "type": "string",
            "enum": [
              "LOW",
              "MODERATE",
              "HIGH",
              "VERY_HIGH"
            ],
            "title": "Risk Category"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "feature_version": {
            "type": "string",
            "title": "Feature Version"
          },
          "threshold_version": {
            "type": "string",
            "title": "Threshold Version"
          },
          "data_completeness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Data Completeness",
            "default": 1.0
          },
          "input_quality": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "BEST",
                  "GOOD",
                  "FAIR",
                  "POOR"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Quality"
          },
          "missing_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Features"
          },
          "forecast_uncertainty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Uncertainty"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "NOWCAST",
              "FORECAST"
            ],
            "title": "Source Type"
          },
          "in_distribution": {
            "type": "boolean",
            "title": "In Distribution",
            "default": true
          },
          "ood_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ood Score"
          },
          "prediction_status": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "PARTIAL",
              "FAILED"
            ],
            "title": "Prediction Status",
            "default": "SUCCESS"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "provenance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "prediction_id",
          "location_id",
          "prediction_time",
          "target_date",
          "risk_probability",
          "risk_category",
          "model_id",
          "model_version",
          "feature_version",
          "threshold_version",
          "source_type"
        ],
        "title": "RiskPredictionOutput",
        "description": "Health risk prediction output."
      },
      "RiskTrainingRequest": {
        "properties": {
          "config": {
            "title": "Config"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "config"
        ],
        "title": "RiskTrainingRequest",
        "description": "Request to train a new risk model."
      },
...
    }
  }
}